+++ crd-generator/api/src/test/java/io/fabric8/crd/generator/CRDGeneratorTest.java
@@ -21,6 +21,7 @@ import io.fabric8.crd.example.basic.BasicStatus;
 import io.fabric8.crd.example.complex.Complex;
 import io.fabric8.crd.example.cyclic.Cyclic;
 import io.fabric8.crd.example.cyclic.CyclicList;
+import io.fabric8.crd.example.deprecated.v2.DeprecationExample;
 import io.fabric8.crd.example.inherited.BaseSpec;
 import io.fabric8.crd.example.inherited.BaseStatus;
 import io.fabric8.crd.example.inherited.Child;
@@ -50,6 +51,7 @@ import io.fabric8.kubernetes.api.model.apiextensions.v1.JSONSchemaProps;
 import io.fabric8.kubernetes.client.CustomResource;
 import io.fabric8.kubernetes.client.utils.Serialization;
 import io.fabric8.kubernetes.model.Scope;
+import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
 import org.opentest4j.AssertionFailedError;
 import org.slf4j.Logger;
@@ -218,6 +220,53 @@ class CRDGeneratorTest {
     final Set<String> dependentClassNames = infos.get(specVersion).getDependentClassNames();
     Arrays.stream(mustContainTraversedClasses).map(Class::getCanonicalName)
         .forEach(c -> assertTrue(dependentClassNames.contains(c), "should contain " + c));
+  }
+
+  @Test
+  void checkDeprecated() {
+    CRDGenerator generator = newCRDGenerator();
+    final String specVersion = "v1";
+    final CRDGenerationInfo info = generator
+        .customResourceClasses(
+            io.fabric8.crd.example.deprecated.v1beta1.DeprecationExample.class,
+            io.fabric8.crd.example.deprecated.v1.DeprecationExample.class,
+            DeprecationExample.class)
+        .forCRDVersions(specVersion)
+        .withOutput(output)
+        .detailedGenerate();
+
+    assertEquals(1, info.numberOfGeneratedCRDs());
+    final Map<String, Map<String, CRDInfo>> details = info.getCRDDetailsPerNameAndVersion();
+    assertEquals(1, details.size());
+    // check multiple versions for same CR
+    final String crdName = CustomResource.getCRDName(DeprecationExample.class);
+    assertTrue(details.containsKey(crdName));
+    final Map<String, CRDInfo> infos = info.getCRDInfos(crdName);
+    assertEquals(1, infos.size());
+    assertTrue(infos.containsKey(specVersion));
+
+    final String outputName = CRDGenerator.getOutputName(crdName, specVersion);
+    CustomResourceDefinition definition = output.definition(outputName);
+    assertNotNull(definition);
+    assertEquals("apiextensions.k8s.io/" + specVersion, definition.getApiVersion());
+
+    CustomResourceDefinitionSpec spec = definition.getSpec();
+    final List<CustomResourceDefinitionVersion> versions = spec.getVersions();
+    assertEquals(3, versions.size());
+    assertEquals(1, versions.stream().filter(v -> v.getName().equals("v1beta1")).count());
+    assertEquals(1, versions.stream().filter(v -> v.getName().equals("v1")).count());
+    assertEquals(1, versions.stream().filter(v -> v.getName().equals("v2")).count());
+
+    CustomResourceDefinitionVersion v1Beta1 = versions.stream().filter(v -> v.getName().equals("v1beta1")).findFirst().get();
+    CustomResourceDefinitionVersion v1 = versions.stream().filter(v -> v.getName().equals("v1")).findFirst().get();
+    CustomResourceDefinitionVersion v2 = versions.stream().filter(v -> v.getName().equals("v2")).findFirst().get();
+    assertTrue(v1Beta1.getDeprecated());
+    assertEquals("sample.fabric8.io/v1beta1 DeprecationExample is deprecated; Migrate to sample.fabric8.io/v2",
+        v1Beta1.getDeprecationWarning());
+    assertTrue(v1.getDeprecated());
+    assertNull(v1.getDeprecationWarning());
+    assertNull(v2.getDeprecated());
+    assertNull(v2.getDeprecationWarning());
   }
 
   @Test
@@ -434,19 +483,65 @@ class CRDGeneratorTest {
   void checkGenerationIsDeterministic() throws Exception {
     // generated CRD
     final File outputDir = Files.createTempDirectory("crd-").toFile();
-    final CustomResourceInfo info = CustomResourceInfo.fromClass(Complex.class);
-    final CRDGenerationInfo crdInfo = newCRDGenerator().inOutputDir(outputDir).forCRDVersions(info.version())
-        .customResources(info).customResourceClasses(Complex.class).detailedGenerate();
-    final File crdFile = new File(crdInfo.getCRDInfos(info.crdName()).get(info.version()).getFilePath());
+    final String crdName = CustomResourceInfo.fromClass(Complex.class).crdName();
+    final CRDGenerationInfo crdInfo = newCRDGenerator()
+        .inOutputDir(outputDir)
+        .forCRDVersions("v1", "v1beta1")
+        .customResourceClasses(Complex.class)
+        .detailedGenerate();
+    final File crdFile = new File(crdInfo.getCRDInfos(crdName).get("v1").getFilePath());
+    final File crdFileV1Beta1 = new File(crdInfo.getCRDInfos(crdName).get("v1beta1").getFilePath());
 
     // expected CRD
     final URL crdResource = CRDGeneratorTest.class.getResource("/" + crdFile.getName());
+    final URL crdResourceV1Beta1 = CRDGeneratorTest.class.getResource("/" + crdFileV1Beta1.getName());
+
     assertNotNull(crdResource);
+    assertNotNull(crdResourceV1Beta1);
     final File expectedCrdFile = new File(crdResource.getFile());
+    final File expectedCrdFileV1Beta1 = new File(crdResourceV1Beta1.getFile());
     assertFileEquals(expectedCrdFile, crdFile);
+    assertFileEquals(expectedCrdFileV1Beta1, crdFileV1Beta1);
 
     // only delete the generated files if the test is successful
     assertTrue(crdFile.delete());
+    assertTrue(crdFileV1Beta1.delete());
+    assertTrue(outputDir.delete());
+  }
+
+  @RepeatedTest(value = 10)
+  void checkGenerationMultipleVersionsOfCRDsIsDeterministic() throws Exception {
+    // generated CRD
+    final File outputDir = Files.createTempDirectory("crd-").toFile();
+    final CustomResourceInfo infoV1 = CustomResourceInfo.fromClass(Multiple.class);
+    final CustomResourceInfo infoV2 = CustomResourceInfo.fromClass(io.fabric8.crd.example.multiple.v2.Multiple.class);
+    assertEquals(infoV1.crdName(), infoV2.crdName());
+    final String crdName = infoV1.crdName();
+
+    final CRDGenerationInfo crdInfo = newCRDGenerator()
+        .inOutputDir(outputDir)
+        .customResourceClasses(Multiple.class,
+            io.fabric8.crd.example.multiple.v2.Multiple.class)
+        .forCRDVersions("v1", "v1beta1")
+        .detailedGenerate();
+
+    final File crdFile = new File(crdInfo.getCRDInfos(crdName).get("v1").getFilePath());
+    final File crdFileV1Beta1 = new File(crdInfo.getCRDInfos(crdName).get("v1beta1").getFilePath());
+
+    // expected CRD
+    final URL crdResource = CRDGeneratorTest.class.getResource("/" + crdFile.getName());
+    final URL crdResourceV1Beta1 = CRDGeneratorTest.class.getResource("/" + crdFileV1Beta1.getName());
+    assertNotNull(crdResource);
+    assertNotNull(crdResourceV1Beta1);
+
+    final File expectedCrdFile = new File(crdResource.getFile());
+    final File expectedCrdFileV1Beta1 = new File(crdResourceV1Beta1.getFile());
+    assertFileEquals(expectedCrdFile, crdFile);
+    assertFileEquals(expectedCrdFileV1Beta1, crdFileV1Beta1);
+
+    // only delete the generated files if the test is successful
+    assertTrue(crdFile.delete());
+    assertTrue(crdFileV1Beta1.delete());
     assertTrue(outputDir.delete());
   }